Set up the header/entry logic for compile-time dispatch.
There are two implementations of the header. One will be fully safe
rust, making it easily auditable. The other is unsafe, but allows for
a smaller, one-word header, which would be necessary for some collection
algorithms (such as mark-compact).
